requirements.txt
beautifulsoup4==4.12.3
click==8.1.7
# git+https://github.com/rossant/ipycache.git # for %%cache. 
dask==2023.5.0 
fcd_torch==1.0.7
gputil==1.4.0
ipykernel==6.29.5
ipywidgets==8.1.5
numpy==1.24.4 # most recent version of numpy that works with python 3.8.10
matplotlib==3.7.1
pandas==1.5.3
pathlib==1.0.1
pyarrow ==17.0.0
ray==2.10.0 # might need to be ray[tune] instead of just ray
rdkit==2023.9.5
requests==2.31.0
scikit-learn==1.2.2
seaborn==0.13.1
# torchmetrics==1.3.1 # make sure these are correct
# torch==2.1.0
# torchvision==0.16.0
wandb==0.17.0